# tests/STLCompliance.OpenApi.Tests/snapshots/routarr.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "RoutArr.Api | v1"
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
  "paths": {
...
    "/api/audit-packages/export": {
      "get": {
        "tags": [
          "AuditPackages"
        ],
        "operationId": "ExportRoutArrAuditPackage",
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "action",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "result",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "targetType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actorUserId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/audit-packages/jobs": {
      "post": {
        "tags": [
          "AuditPackages"
        ],
        "operationId": "CreateRoutArrAuditPackageGenerationJob",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAuditPackageGenerationJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/audit-packages/jobs/{jobId}": {
      "get": {
        "tags": [
          "AuditPackages"
        ],
        "operationId": "GetRoutArrAuditPackageGenerationJob",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/audit-packages/jobs/{jobId}/download": {
      "get": {
        "tags": [
          "AuditPackages"
        ],
        "operationId": "DownloadRoutArrAuditPackageGenerationJob",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/audit-packages/summary": {
      "get": {
        "tags": [
          "AuditPackages"
        ],
        "operationId": "GetRoutArrAuditPackageExportSummary",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "action",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "result",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "targetType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actorUserId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/audit-packages/timeline": {
      "get": {
        "tags": [
          "AuditPackages"
        ],
        "operationId": "GetRoutArrAuditPackageTimeline",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "action",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "result",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "targetType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actorUserId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateAuditPackageGenerationJobRequest": {
        "required": [
          "format",
          "from",
          "to"
        ],
        "type": "object",
        "properties": {
          "format": {
            "type": "string"
          },
          "from": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "to": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "action": {
            "type": [
              "null",
              "string"
            ]
          },
          "result": {
            "type": [
              "null",
              "string"
            ]
          },
          "targetType": {
            "type": [
              "null",
              "string"
            ]
          },
          "actorUserId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
...
    }
  },
...
}